Move Page.copy_for_translation ino an external function
