+++ wagtail/core/models/__init__.py
@@ -46,6 +46,7 @@ from modelcluster.fields import ParentalKey
 from modelcluster.models import ClusterableModel
 from treebeard.mp_tree import MP_Node
 
+from wagtail.core.actions.copy_for_translation import copy_for_translation as _copy_for_translation
 from wagtail.core.actions.copy_page import CopyPageAction
 from wagtail.core.actions.delete_page import DeletePageAction
 from wagtail.core.actions.move_page import MovePageAction
@@ -61,7 +62,7 @@ from wagtail.core.signals import (
 from wagtail.core.treebeard import TreebeardPathFixMixin
 from wagtail.core.url_routing import RouteResult
 from wagtail.core.utils import (
-    WAGTAIL_APPEND_SLASH, camelcase_to_underscore, find_available_slug,
+    WAGTAIL_APPEND_SLASH, camelcase_to_underscore,
     get_supported_content_language_variant, resolve_model_string)
 from wagtail.search import index
 
@@ -101,14 +102,6 @@ def reassign_root_page_locale_on_delete(sender, instance, **kwargs):
             new_locale = Locale.all_objects.exclude(pk=instance.pk).first()
 
         root_page_with_this_locale.update(locale=new_locale)
-
-
-class ParentNotTranslatedError(Exception):
-    """
-    Raised when a call to Page.copy_for_translation is made but the
-    parent page is not translated and copy_parents is False.
-    """
-    pass
 
 
 PAGE_MODEL_CLASSES = []
@@ -1602,7 +1595,6 @@ class Page(AbstractPage, index.Indexed, ClusterableModel, metaclass=PageBase):
 
     create_alias.alters_data = True
 
-    @transaction.atomic
     def copy_for_translation(self, locale, copy_parents=False, alias=False, exclude_fields=None):
         """
         Creates a copy of this page in the specified locale.
@@ -1624,60 +1616,8 @@ class Page(AbstractPage, index.Indexed, ClusterableModel, metaclass=PageBase):
         Note that this method calls the ``.copy()`` method internally so any fields that
         are excluded in ``.exclude_fields_in_copy`` will be excluded from the translation.
         """
-        # Find the translated version of the parent page to create the new page under
-        parent = self.get_parent().specific
-        slug = self.slug
 
-        if not parent.is_root():
-            try:
-                translated_parent = parent.get_translation(locale)
-            except parent.__class__.DoesNotExist:
-                if not copy_parents:
-                    raise ParentNotTranslatedError
-
-                translated_parent = parent.copy_for_translation(
-                    locale, copy_parents=True, alias=True
-                )
-        else:
-            # Don't duplicate the root page for translation. Create new locale as a sibling
-            translated_parent = parent
-
-            # Append language code to slug as the new page
-            # will be created in the same section as the existing one
-            slug += "-" + locale.language_code
-
-        # Find available slug for new page
-        slug = find_available_slug(translated_parent, slug)
-
-        if alias:
-            return self.create_alias(
-                parent=translated_parent,
-                update_slug=slug,
-                update_locale=locale,
-                reset_translation_key=False,
-            )
-
-        else:
-            # Update locale on translatable child objects as well
-            def process_child_object(
-                original_page, page_copy, child_relation, child_object
-            ):
-                if isinstance(child_object, TranslatableMixin):
-                    child_object.locale = locale
-
-            return self.copy(
-                to=translated_parent,
-                update_attrs={
-                    "locale": locale,
-                    "slug": slug,
-                },
-                copy_revisions=False,
-                keep_live=False,
-                reset_translation_key=False,
-                process_child_object=process_child_object,
-                exclude_fields=exclude_fields,
-                log_action='wagtail.copy_for_translation',
-            )
+        return _copy_for_translation(self, locale, copy_parents=copy_parents, alias=alias, exclude_fields=exclude_fields)
 
     copy_for_translation.alters_data = True
